+++ gulp/res/js/yous.js
@@ -5,7 +5,19 @@ let yousEnabled = localStorage.getItem('yous-setting') == 'true';
 let savedYous = new Set(JSON.parse(localStorage.getItem('yous')));
 let yousList;
 
-const toggleAllYous = (state) => savedYous.forEach(y => toggleOne(y, state));
+function clearYousList() {
+	if (yousEnabled) {
+		toggleAllYous(false);
+	}
+	savedYous = new Set();
+	yousList.value = '';
+	setLocalStorage('yous', '[]');
+	console.log('cleared yous');
+}
+
+function toggleAllYous(state) {
+	savedYous.forEach(y => toggleOne(y, state));
+}
 
 const toggleQuotes = (quotes, state) => {
 	quotes.forEach(q => {
@@ -63,6 +75,26 @@ if (yousEnabled) {
 	toggleAllYous(yousEnabled);
 }
 
+//handle when (you)s change in a different tab
+const yousStorageEventHandler = (e) => {
+	if (e.storageArea === localStorage
+		&& e.key === 'yous') {
+		const newArrayYous = [...JSON.parse(e.newValue)];
+		newArrayYous.forEach(ny => {
+			if (!savedYous.has(ny)) {
+				savedYous.add(ny);
+				console.log('handle new (you) from other context:', ny);
+				toggleOne(ny, yousEnabled);
+			}
+		});
+		if (newArrayYous.length === 0) {
+			console.log('(you)s were cleared in another tab');
+			clearYousList();
+		}
+		yousList.value = newArrayYous.toString();
+	}
+};
+
 const handleNewYous = (e) => {
 	const postYou = `${e.detail.json.board}-${e.detail.postId}`;
 	const isYou = window.myPostId == e.detail.postId;
@@ -111,6 +143,7 @@ const handleNewYous = (e) => {
 };
 
 window.addEventListener('addPost', handleNewYous, false);
+window.addEventListener('storage', yousStorageEventHandler, false);
 window.addEventListener('updatePostMessage', handleNewYous, false);
 
 window.addEventListener('settingsReady', () => {
@@ -118,15 +151,6 @@ window.addEventListener('settingsReady', () => {
 	yousList = document.getElementById('youslist-setting');
 	yousList.value = [...savedYous];
 	const yousListClearButton = document.getElementById('youslist-clear');
-	const clearYousList = () => {
-		if (yousEnabled) {
-			toggleAllYous(false);
-		}
-		savedYous = new Set();
-		yousList.value = '';
-		setLocalStorage('yous', '[]');
-		console.log('cleared yous');
-	};
 	yousListClearButton.addEventListener('click', clearYousList, false);
 
 	const yousSetting = document.getElementById('yous-setting');